check for trashed items
only apply these functions on trashed items when they are being deleted permanently

+++ app/Events/BlogDeleted.php
@@ -18,8 +18,10 @@ class BlogDeleted
     public function __construct(Blog $blog)
     {
         $this->blog = $blog;
-        // Delet Blog Image
-        Storage::delete($this->blog->image);
+        if ($this->blog->deleted_at != null) {
+            // Delet Blog Image
+            Storage::delete($this->blog->image);
+        }
     }
 
     /**
